Fix errors in SortedJsonFormatter
The tuple being passed through an ignore-arg-type was actually
incorrect and this wasn't being examined. Secondarily,
parse_null_values needed to be set to `True` for `None -> null` to
happen as expected.

In service of this, add a new test dir for formatter tests and move
existing tests into that dir. The new tests for SortedJsonFormatter
reside in the new dir.
