+++ tests/unit/formatters/test_primitive_formatters.py
@@ -1,0 +1,29 @@
+import pytest
+
+from globus_cli.termio import Field, formatters
+
+
+def test_date_format():
+    f = Field("foo", "foo", formatter=formatters.Date)
+    assert f({"foo": None}) == "None"
+    assert f({"foo": "2022-04-05T16:27:48.805427"}) == "2022-04-05 16:27:48"
+
+
+def test_bool_format():
+    f = Field("foo", "foo", formatter=formatters.Bool)
+    assert f({"foo": None}) == "None"
+    assert f({}) == "None"
+    assert f({"foo": True}) == "True"
+    assert f({"foo": False}) == "False"
+
+    with pytest.warns(formatters.FormattingFailedWarning):
+        assert f({"foo": "hi there"}) == "hi there"
+
+
+def test_fuzzy_bool_format():
+    f = Field("foo", "foo", formatter=formatters.FuzzyBool)
+    assert f({"foo": None}) == "False"
+    assert f({}) == "False"
+    assert f({"foo": True}) == "True"
+    assert f({"foo": False}) == "False"
+    assert f({"foo": "hi there"}) == "True"
